Renaming the package io.adminshell.aas.v3.dataformat.json-ld to io.adminshell.aas.v3.dataformat.rdf
